@@ -7,45 +7,46 @@
     "test": "cross-env NODE_ENV=testing jest --watchAll --verbose --forceExit"
   },
   "devDependencies": {
-    "@babel/core": "7.17.5",
-    "@babel/plugin-transform-react-jsx": "7.17.3",
-    "@babel/plugin-transform-runtime": "7.17.0",
-    "@babel/preset-env": "7.16.11",
-    "@babel/preset-react": "7.16.7",
-    "@testing-library/jest-dom": "5.16.2",
-    "@testing-library/react": "12.1.3",
-    "@types/jest": "27.4.0",
-    "babel-loader": "8.2.3",
-    "babel-plugin-styled-components": "2.0.3",
-    "concurrently": "7.0.0",
+    "@babel/core": "7.18.10",
+    "@babel/plugin-transform-react-jsx": "7.18.10",
+    "@babel/plugin-transform-runtime": "7.18.10",
+    "@babel/preset-env": "7.18.10",
+    "@babel/preset-react": "7.18.6",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.3.0",
+    "@types/jest": "28.1.7",
+    "babel-loader": "8.2.5",
+    "babel-plugin-styled-components": "2.0.7",
+    "concurrently": "7.3.0",
     "cross-env": "7.0.3",
-    "eslint": "8.9.0",
-    "eslint-plugin-react": "7.28.0",
+    "eslint": "8.22.0",
+    "eslint-plugin-react": "7.30.1",
     "fkill-cli": "7.1.0",
-    "html-loader": "3.1.0",
+    "html-loader": "4.1.0",
     "html-webpack-plugin": "5.5.0",
-    "jest": "27.5.1",
-    "msw": "0.38.1",
-    "nodemon": "2.0.15",
-    "webpack": "5.69.1",
-    "webpack-dev-server": "4.7.4"
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "msw": "0.45.0",
+    "nodemon": "2.0.19",
+    "webpack": "5.74.0",
+    "webpack-dev-server": "4.10.0"
   },
   "dependencies": {
-    "@testing-library/user-event": "13.5.0",
-    "axios": "0.26.0",
+    "@testing-library/user-event": "14.4.3",
+    "axios": "0.27.2",
     "cors": "2.8.5",
-    "css-loader": "6.6.0",
-    "express": "4.17.3",
-    "moment": "2.29.1",
+    "css-loader": "6.7.1",
+    "express": "4.18.1",
+    "moment": "2.29.4",
     "mutationobserver-shim": "0.3.7",
-    "nanoid": "3.3.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "6.2.1",
+    "nanoid": "4.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.3.0",
     "striptags": "3.2.0",
     "style-loader": "3.3.1",
-    "styled-components": "5.3.3",
-    "webpack-cli": "4.9.2"
+    "styled-components": "5.3.5",
+    "webpack-cli": "4.10.0"
   },
   "repository": {
     "type": "git",
